feat: Dice number slider value tracing the graph anew
The slider value is now gotten and causes the graph to be redrawn.
The good news is that, for big enough values, the graph does present a
peak, which intuitively makes sense. I'm glad I've seen this behavior as
I was expecting it, but have not written explicit tests for it
beforehand.