refacto(stats): More logical way to express the stats function
After talking about it with a colleague, I've decided to change the way
the function is formulated. This way makes more sense to me:
proba = proba of getting k matches * proba of getting n-k non-matches *
the binomial coefficient.
This will also make it easier to use the same function to calculate the
probability of getting matches with the Paco rule later on.