fix(stats): Removal of the posterior proba function
After reviewing the calculations, I made a mistake and this function is
not useful. The reason for that is that, when expressing Bayes' rule,
most terms simplify and in the end, the posterior probability
ends up being the same expression as calculating the probability of
getting a given face with only the other's dices.
Thinking about it, I suppose this makes sense given all the dice rolls
are independant.